refactor: dry up scoped symbols
