+++ pkgs/sass_language_services/lib/sass_language_services.dart
@@ -23,4 +23,5 @@ export 'src/file_system_provider.dart'
 export 'src/language_services.dart' show LanguageServices;
 
 export 'src/features/document_links/stylesheet_document_link.dart';
+export 'src/features/document_symbols/scoped_symbols.dart';
 export 'src/features/document_symbols/stylesheet_document_symbol.dart';
