+++ pkgs/sass_language_services/lib/src/features/document_symbols/document_symbols_feature.dart
@@ -14,4 +14,17 @@ class DocumentSymbolsFeature extends LanguageFeature {
 
     return symbolsVisitor.symbols;
   }
+
+  ScopedSymbols getScopedSymbols(TextDocument document) {
+    var stylesheet = ls.parseStylesheet(document);
+
+    var symbols = ls.cache.getScopedSymbols(document) ??
+        ScopedSymbols(
+          stylesheet,
+          document.languageId == 'sass' ? Dialect.indented : Dialect.scss,
+        );
+    ls.cache.setScopedSymbols(document, symbols);
+
+    return symbols;
+  }
 }
