+++ pkgs/sass_language_services/lib/src/features/document_symbols/scope.dart
@@ -1,4 +1,4 @@
-import '../document_symbols/stylesheet_document_symbol.dart';
+import 'stylesheet_document_symbol.dart';
 
 class Scope {
   Scope? parent;
