+++ pkgs/sass_language_services/test/features/go_to_definition/scoped_symbols_test.dart
@@ -1,6 +1,4 @@
 import 'package:sass_language_services/sass_language_services.dart';
-import 'package:sass_language_services/src/features/go_to_definition/scope_visitor.dart';
-import 'package:sass_language_services/src/features/go_to_definition/scoped_symbols.dart';
 import 'package:test/test.dart';
 
 import '../../memory_file_system.dart';
